requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.12
pydantic>=2.9.0
web3==6.11.1
opencv-python-headless>=4.10.0.84
numpy>=2.1.0
PyPDF2>=3.0.1
librosa>=0.10.2
soundfile>=0.12.1
python-dotenv>=1.0.1
requests>=2.32.0
supabase>=2.10.0
reportlab>=4.2.5
face-recognition>=1.3.0
